+++ src/lexer/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./common";
-export * from "./numeric";
 export * from "./string";
 export * from "./word";
 export * as javascript from "./javascript";