+++ src/lexer/utils/javascript/index.ts
@@ -1,3 +1,4 @@
 export * from "./common";
 export * from "./string";
 export * from "./regex";
+export * from "./numeric";